requirements.txt
Django==3.2.6
mysqlclient==2.0.3
xlrd
pandas==1.3.4
PyMySQL
djangorestframework
pytest-django